# docs/runtime-api.full.openapi.yaml
...
openapi: 3.0.3
info:
  title: MisterMorph Runtime API — Full
  version: '2026-10-02'
  description: Full Runtime API reference for shared control/data routes and runtime-mounted management,
    provider login and setup extensions. Availability varies by runtime; Console-specific operations are
    marked. Browser-only /api routes are outside this runtime API. The document revision is not a negotiated
    protocol version. Use runtime-api.control.openapi.yaml for task execution, results and approvals.
    See runtime-api.md for behavior and examples.
servers:
- url: http://127.0.0.1:9080/runtime
  description: Default Console listener after setting server.auth_token. Replace with the full runtime
    base URL, including any console.base_path or reverse-proxy prefix.
security:
- RuntimeBearer: []
...
paths:
...
  /overview:
    get:
      operationId: getOverview
      tags:
      - Runtime
      summary: Read runtime identity, counters and process metrics
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Overview'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '503':
          $ref: '#/components/responses/Unavailable'
  /commands:
    get:
      operationId: listCommands
      tags:
      - Runtime
      summary: List slash command suggestions
      description: Submit a command as task text to POST /tasks. This catalog is shared; individual runtimes
        can support different commands.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CommandSuggestion'
                required:
                - items
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /llm/profiles:
    get:
      operationId: listLLMProfiles
      tags:
      - Runtime
      summary: Read default LLM route and named profiles
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMProfiles'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /poke:
    post:
      operationId: pokeRuntime
      tags:
      - Runtime
      summary: Trigger awareness with external text
      description: Requires a poke handler. Returns 409 when awareness is already running. This does not
        return a task ID; use POST /tasks when a task result is needed.
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              minLength: 1
        description: Nonblank textual body, at most 10 KiB in bytes. JSON/XML/text media types are also
          accepted as text; there is no JSON task envelope.
      responses:
        '202':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  mode:
                    type: string
                  poked_at:
                    type: string
                    format: date-time
                required:
                - ok
                - mode
                - poked_at
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '503':
          $ref: '#/components/responses/Unavailable'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/TooLarge'
...
  /todo/tasks/{id}/run:
    post:
      operationId: runScheduledTask
      tags:
      - Schedules
      summary: Trigger a saved scheduled task now
      description: No request body. Requires a cron runner. A busy runner returns 409; a missing saved
        schedule returns 404.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      responses:
        '202':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  id:
                    type: string
                  triggered_at:
                    type: string
                    format: date-time
                required:
                - ok
                - id
                - triggered_at
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '503':
          $ref: '#/components/responses/Unavailable'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /stats/llm/usage:
    get:
      operationId: getLLMUsage
      tags:
      - Observability
      summary: Read aggregate LLM usage
      description: Summary and per-host/model entries are extensible llmstats aggregate payloads.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  generated_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                  projected_offset:
                    type: object
                    additionalProperties: true
                  projected_records:
                    type: integer
                  skipped_records:
                    type: integer
                  summary:
                    type: object
                    additionalProperties: true
                  api_hosts:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
                    nullable: true
                  models:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
                    nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
  /stats/llm/daily:
    get:
      operationId: getDailyLLMUsage
      tags:
      - Observability
      summary: Read daily LLM usage
      description: Daily rows are oldest first, including days with zero usage. Summary and daily model
        counters are extensible llmstats payloads.
      parameters:
      - name: days
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 30
        description: Values above 366 are clamped to 366.
      - name: tz
        in: query
        required: false
        schema:
          type: string
          default: UTC
        description: IANA time zone, for example Asia/Tokyo.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  generated_at:
                    type: string
                    format: date-time
                  time_zone:
                    type: string
                  from:
                    type: string
                    format: date
                  to:
                    type: string
                    format: date
                  summary:
                    type: object
                    additionalProperties: true
                  models:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
                    nullable: true
                  days:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
  /system/diagnostics:
    get:
      operationId: getDiagnostics
      tags:
      - Observability
      summary: Read runtime filesystem diagnostics
      description: Checks readability and writability of configured server paths.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  started_at:
                    type: string
                    format: date-time
                  version:
                    type: string
                  checks:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        ok:
                          type: boolean
                        detail:
                          type: string
                required:
                - started_at
                - version
                - checks
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
...
  /settings/system:
    get:
      operationId: getSystemSettings
      tags:
      - Management
      summary: Read System settings
      description: Console only.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettingsView'
        '401':
          description: Missing or invalid runtime token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONError'
        '405':
          description: Unsupported method or read-only settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONError'
        '500':
          description: Storage or internal failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONError'
    put:
      operationId: updateSystemSettings
      tags:
      - Management
      summary: Update System settings
      description: Console only. JSON body limit is 1 MiB. A nonempty config_revision must match the latest
        revision. Apply status and restart requirements are returned in the response.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettingsView'
        '400':
          description: Invalid request or configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONError'
        '401':
          description: Missing or invalid runtime token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONError'
        '405':
          description: Unsupported method or read-only settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONError'
        '409':
          description: Concurrent configuration change or operation conflicts with current state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONError'
        '500':
          description: Storage or internal failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONError'
        '502':
          description: Remote provider or endpoint request failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemSettingsUpdate'
...
components:
  securitySchemes:
    RuntimeBearer:
      type: http
      scheme: bearer
      description: The configured server.auth_token. This is not a Console browser session token.
  responses:
    BadRequest:
      description: Invalid JSON, parameter or runtime request.
      content:
        text/plain:
          schema:
            type: string
    Unauthorized:
      description: Missing or invalid Authorization Bearer token.
      content:
        text/plain:
          schema:
            type: string
    NotFound:
      description: Resource or route not found.
      content:
        text/plain:
          schema:
            type: string
    MethodNotAllowed:
      description: Unsupported method; task submission can also be disabled.
      content:
        text/plain:
          schema:
            type: string
    Conflict:
      description: Concurrent change or runtime busy.
      content:
        text/plain:
          schema:
            type: string
    TooLarge:
      description: Request exceeds the endpoint byte limit.
      content:
        text/plain:
          schema:
            type: string
    InternalError:
      description: Runtime storage or internal failure.
      content:
        text/plain:
          schema:
            type: string
    Unavailable:
      description: Required runtime capability is absent, queue is full, or execution/storage is unavailable.
      content:
        text/plain:
          schema:
            type: string
  schemas:
...
    Overview:
      type: object
      properties:
        health:
          type: string
        mode:
          type: string
        agent_name:
          type: string
        agent_avatar_url:
          type: string
        submit_enabled:
          type: boolean
        instance_id:
          type: string
        started_at:
          type: string
          format: date-time
        uptime_sec:
          type: integer
        last_poke_at:
          type: string
          format: date-time
        version:
          type: string
        runtime:
          type: object
          additionalProperties: true
      description: Additional runtime-specific counters and channel data may be present.
    CommandSuggestion:
      type: object
      properties:
        value:
          type: string
        title:
          type: string
        description:
          type: string
        insert_text:
          type: string
      required:
      - value
      - title
      - insert_text
      x-source: internal/runtimecommands/suggestions.go:Suggestion
    LLMProfile:
      type: object
      properties:
        name:
          type: string
        inference_provider:
          type: string
        model:
          type: string
      required:
      - name
      x-source: internal/daemonruntime/server.go:runtimeLLMProfileOption
    LLMProfiles:
      type: object
      properties:
        default:
          $ref: '#/components/schemas/LLMProfile'
        items:
          type: array
          items:
            $ref: '#/components/schemas/LLMProfile'
      required:
      - default
      - items
...
    JSONError:
      type: object
      properties:
        ok:
          type: boolean
        error:
          type: string
      required:
      - error
...
    SystemSettingsView:
      type: object
      properties:
        ok:
          type: boolean
        config_path:
          type: string
        config_revision:
          type: string
        config_values:
          type: object
          additionalProperties: true
        field_states:
          type: object
          additionalProperties: true
        apply_mode:
          type: string
        apply_status:
          type: string
        restart_targets:
          type: array
          items:
            type: string
          nullable: true
    SystemSettingsUpdate:
      type: object
      properties:
        config_revision:
          type: string
        config_changes:
          type: object
          additionalProperties: true
        reset:
          type: array
          items:
            type: string
          nullable: true
      description: At least one config_changes or reset entry is required. config_changes uses registered
        dotted system-setting keys. A nonempty stale config_revision returns 409.
...